代码提交

+++ iOS-Objc-Demo/2018-09/2018-09-14/自定义图片选择/WDPhotoPickerTools/Controller/WDPhotoPickerViewController.m
@@ -160,7 +160,7 @@
             self.bottomView.sureButton.backgroundColor = UIColor.lightGrayColor;
             self.bottomView.sureButton.userInteractionEnabled = false;
         } else {
-            [self.bottomView.sureButton setTitle:[NSString stringWithFormat:@"确定(%ld)",self.selectedArray.count] forState:UIControlStateNormal];
+            [self.bottomView.sureButton setTitle:[NSString stringWithFormat:@"确定(%lu)",(unsigned long)self.selectedArray.count] forState:UIControlStateNormal];
             self.bottomView.sureButton.backgroundColor = self.sureBackgrondColor;
             self.bottomView.sureButton.userInteractionEnabled = true;
         }
